extracts mock streams and test data helpers
