+++ tests/common/mod.rs
@@ -1,0 +1,7 @@
+//! Common test utilities and helper modules
+//! 
+//! This module provides shared functionality for all test files,
+//! including mock streams and test data creation utilities.
+
+pub mod mock_streams;
+pub mod test_data;
